requirements.txt
pandas
numpy
sqlalchemy
psycopg2-binary
requests
python-dotenv
joblib
scikit-learn
lightgbm
ta
websocket-client
shap
